+++ test/ash_onetime/action_contention_test.exs
@@ -2,17 +2,9 @@ defmodule AshOnetime.ActionContentionTest do
   use ExUnit.Case, async: false
 
   alias AshOnetime.Test.ActionExamples.Resource
-  alias AshOnetime.Test.{Migration, RealConnection, Repo}
+  alias AshOnetime.Test.{ExternalPeer, Migration, RealConnection, Repo}
   alias Ecto.Adapters.SQL
   alias Ecto.Adapters.SQL.Sandbox
-
-  @database_options [
-    hostname: "127.0.0.1",
-    port: 18_841,
-    username: "postgres",
-    password: "postgres",
-    database: "ash_onetime_test"
-  ]
 
   setup_all do
     installation = Migration.install_generated!()
@@ -120,7 +112,7 @@ defmodule AshOnetime.ActionContentionTest do
   end
 
   defp observer! do
-    {:ok, observer} = Postgrex.start_link(@database_options)
+    {:ok, observer} = Postgrex.start_link(ExternalPeer.database_options())
     Process.unlink(observer)
 
     on_exit(fn ->