+++ test/mix/tasks/ash_onetime.gen.migrations_test.exs
@@ -1,3 +1,5 @@
+Code.require_file("../../../scripts/portable.exs", __DIR__)
+
 defmodule Mix.Tasks.AshOnetime.Gen.MigrationsTest do
   use ExUnit.Case, async: false
 
@@ -78,7 +80,7 @@ defmodule Mix.Tasks.AshOnetime.Gen.MigrationsTest do
     package = Path.join(temporary, "package")
 
     {output, status} =
-      System.cmd("mix", ["hex.build", "--unpack", "--output", package],
+      AshOnetime.Portable.cmd("mix", ["hex.build", "--unpack", "--output", package],
         cd: File.cwd!(),
         stderr_to_stdout: true
       )